+++ src/index.js
@@ -212,22 +212,6 @@ const customBlocks = {
       view: [],
     },
   },
-  buttonBlock: {
-    availableColors: ['#ebebeb', '#023d6b', '#ddeeff', '#eb5f73', '#b9d25f'],
-    id: 'buttonBlock',
-    title: 'Button',
-    icon: CircleMenuSVG,
-    group: 'common',
-    view: withStyleWrapper(ButtonViewBlock),
-    edit: withStyleWrapper(ButtonEditBlock),
-    restricted: false,
-    mostUsed: true,
-    sidebarTab: 1,
-    security: {
-      addPermission: [],
-      view: [],
-    },
-  },
 };
 
 // export const blocks = {